Reorganize create exam components
