+++ frontend/src/react/Exam/CreateExam/Exercice/ExerciceSimplified/ExerciceSimplified.js
