+++ frontend/src/react/Exam/CreateExam/Instructions/Instructions.js
@@ -1,11 +1,11 @@
 import React, {Component} from 'react'
 import InstructionsDisplayer from './InstructionsDisplayer';
-import {groupsArray,group} from '../../helpers/mocks/group';
+import {groupsArray,group} from '../../../../helpers/mocks/group';
 
 import {Redirect} from 'react-router-dom';
 import connect from 'react-redux/es/connect/connect';
-import {postExam} from '../../redux/exams/actions/post';
-import {triggerActive,triggerInactive} from '../../helpers/css';
+import {postExam} from '../../../../redux/exams/actions/post';
+import {triggerActive,triggerInactive} from '../../../../helpers/css';
 
 class Instructions extends Component {
     state = {
